requirements.txt
streamlit
DECIMER
rdkit
Pillow